pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ask-the-code"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8,<4.0"
license = "MIT"
keywords = []
authors = [
  { name = "Markis Taylor", email = "m@rkis.net" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "chromadb>=0.5.5",
  "click>=8.1.7",
  "dynaconf>=3.2.6",
  "fast-depends>=2.4.3",
  "flagembedding>=1.2.11",
  "gitpython>=3.1.43",
  "huggingface-hub>=0.24.6",
  "marqo>=3.7.0",
  "mistletoe>=1.4.0",
  "ollama>=0.3.2",
  "peft>=0.12.0",
  "platformdirs>=4.2.2",
  "rich>=13.8.0",
  "typing-extensions>=4.12.2",
]

[project.urls]
Documentation = "https://github.com/markis/ask-the-code#readme"
Issues = "https://github.com/markis/ask-the-code/issues"
Source = "https://github.com/markis/ask-the-code"

[project.scripts]
ask = "ask_the_code.cli:run"
ask-the-code = "ask_the_code.cli:run"

[tool.hatch.version]
path = "src/ask_the_code/__about__.py"

[tool.hatch.envs.default]
installer = "uv"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/ask_the_code tests}"

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.coverage.run]
source_pkgs = ["ask_the_code", "tests"]
branch = true
parallel = true
omit = [
  "src/ask_the_code/__about__.py",
]

[tool.coverage.paths]
ask_the_code = ["src/ask_the_code", "*/ask-the-code/src/ask_the_code"]
tests = ["tests", "*/ask-the-code/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.ruff.lint]
ignore = ["B008", "TCH"]
